requirements.txt
pyserial
regex

